feat: refactor rust commands
